requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
fashion-clip
google-genai
python-multipart
pinecone
psycopg2-binary
